Extract GDTF catalog matcher
