+++ tests/mvr_gdtf_import_matching_test.cpp
@@ -1,10 +1,12 @@
+#include "gdtf_catalog_matcher.h"
 #include "gdtf_import_matching.h"
 
 #include <cassert>
 #include <string>
 #include <vector>
 
-namespace matching = mvr::gdtf_import_matching;
+namespace catalog = mvr::gdtf_catalog_matcher;
+namespace import_matching = mvr::gdtf_import_matching;
 
 struct DownloadRequestProbe {
   std::string requestedFixtureName;
@@ -15,162 +17,134 @@ struct DownloadRequestProbe {
 static void VerifyDistinctOriginalNamesBeatSharedPlaceholderMetadata() {
   const std::string placeholderType = "BLED Standard mode 12CH";
   const std::vector<DownloadRequestProbe> requests = {
-      {matching::SelectRequestedFixtureName("Aleda K10 B-EYE",
-                                            "Fixtures/BLED Standard mode 12CH.gdtf"),
+      {import_matching::SelectRequestedFixtureName("Aleda K10 B-EYE",
+                                                   "Fixtures/BLED Standard mode 12CH.gdtf"),
        placeholderType},
-      {matching::SelectRequestedFixtureName("Super Storm 1500",
-                                            "Fixtures/BLED Standard mode 12CH.gdtf"),
+      {import_matching::SelectRequestedFixtureName("Super Storm 1500",
+                                                   "Fixtures/BLED Standard mode 12CH.gdtf"),
        placeholderType},
   };
 
-  assert(matching::SelectDownloadSearchFixtureName(requests[0].requestedFixtureName,
-                                                   requests[0].resolvedFixtureTypeName) ==
+  assert(catalog::SelectDownloadSearchFixtureName(requests[0].requestedFixtureName,
+                                                  requests[0].resolvedFixtureTypeName) ==
          "Aleda K10 B-EYE");
-  assert(matching::SelectDownloadSearchFixtureName(requests[1].requestedFixtureName,
-                                                   requests[1].resolvedFixtureTypeName) ==
+  assert(catalog::SelectDownloadSearchFixtureName(requests[1].requestedFixtureName,
+                                                  requests[1].resolvedFixtureTypeName) ==
          "Super Storm 1500");
-  assert(matching::SelectDownloadSearchFixtureName(requests[0].requestedFixtureName,
-                                                   requests[0].resolvedFixtureTypeName) !=
+  assert(catalog::SelectDownloadSearchFixtureName(requests[0].requestedFixtureName,
+                                                  requests[0].resolvedFixtureTypeName) !=
          placeholderType);
-  assert(matching::SelectDownloadSearchFixtureName(requests[1].requestedFixtureName,
-                                                   requests[1].resolvedFixtureTypeName) !=
+  assert(catalog::SelectDownloadSearchFixtureName(requests[1].requestedFixtureName,
+                                                  requests[1].resolvedFixtureTypeName) !=
          placeholderType);
 }
 
 // Verifies that the GDTFSpec basename is used when the fixture node name is unavailable.
 static void VerifySpecBasenameFallback() {
-  assert(matching::SelectRequestedFixtureName("", "Folder/Super Storm 1500.gdtf") ==
+  assert(import_matching::SelectRequestedFixtureName("", "Folder/Super Storm 1500.gdtf") ==
          "Super Storm 1500");
-  assert(matching::SelectRequestedFixtureName("\t\n", "Folder\\Aleda K10 B-EYE.gdtf") ==
+  assert(import_matching::SelectRequestedFixtureName("\t\n", "Folder\\Aleda K10 B-EYE.gdtf") ==
          "Aleda K10 B-EYE");
 }
 
 // Verifies that resolved metadata remains the final fallback when no original MVR identity exists.
 static void VerifyResolvedTypeFallback() {
-  assert(matching::SelectDownloadSearchFixtureName("", "BLED Standard mode 12CH") ==
+  assert(catalog::SelectDownloadSearchFixtureName("", "BLED Standard mode 12CH") ==
          "BLED Standard mode 12CH");
 }
 
 // Verifies name confidence outranks footprint matches.
 static void VerifyExactNameWithoutFootprintBeatsPartialNameWithFootprint() {
-  const auto exactTier = matching::ComputeFixtureNameMatchTier("My Beam 200",
-                                                              "My Beam 200");
-  const auto partialTier = matching::ComputeFixtureNameMatchTier("My Beam 200 Extended",
-                                                                "My Beam 200");
-  const matching::DownloadCandidateRank exactNoFootprint{
+  const auto exactTier = catalog::ComputeFixtureNameMatchTier("My Beam 200",
+                                                             "My Beam 200");
+  const auto partialTier = catalog::ComputeFixtureNameMatchTier("My Beam 200 Extended",
+                                                               "My Beam 200");
+  const catalog::DownloadCandidateRank exactNoFootprint{
       exactTier, false, false, 100, 0.0f};
-  const matching::DownloadCandidateRank partialWithFootprint{
+  const catalog::DownloadCandidateRank partialWithFootprint{
       partialTier, true, false, 200, 5.0f};
 
-  assert(exactTier == matching::FixtureNameMatchTier::ExactNormalized);
-  assert(partialTier == matching::FixtureNameMatchTier::Partial);
-  assert(matching::IsBetterDownloadCandidate(exactNoFootprint, partialWithFootprint));
-  assert(!matching::IsBetterDownloadCandidate(partialWithFootprint, exactNoFootprint));
+  assert(exactTier == catalog::FixtureNameMatchTier::ExactNormalized);
+  assert(partialTier == catalog::FixtureNameMatchTier::Partial);
+  assert(catalog::IsBetterDownloadCandidate(exactNoFootprint, partialWithFootprint));
+  assert(!catalog::IsBetterDownloadCandidate(partialWithFootprint, exactNoFootprint));
 }
 
 // Verifies manufacturer matches outrank recency ties.
 static void VerifyManufacturerMatchBeatsRecencyInsideSameNameAndFootprintTier() {
-  const auto tier = matching::ComputeFixtureNameMatchTier("My Beam 200", "My Beam 200");
-  const matching::DownloadCandidateRank manufacturerMatch{
+  const auto tier = catalog::ComputeFixtureNameMatchTier("My Beam 200", "My Beam 200");
+  const catalog::DownloadCandidateRank manufacturerMatch{
       tier, true, true, 100, 0.0f};
-  const matching::DownloadCandidateRank newerWithoutManufacturer{
+  const catalog::DownloadCandidateRank newerWithoutManufacturer{
       tier, true, false, 200, 5.0f};
 
-  assert(matching::IsBetterDownloadCandidate(manufacturerMatch,
-                                             newerWithoutManufacturer));
-  assert(!matching::IsBetterDownloadCandidate(newerWithoutManufacturer,
-                                              manufacturerMatch));
-}
-
-struct CatalogModeProbe {
-  std::string name;
-  int footprint = 0;
-};
-
-struct DownloadCandidateProbe {
-  std::string rid;
-  std::string fixtureName;
-  std::vector<CatalogModeProbe> modes;
-  long long recency = 0;
-  float rating = 0.0f;
-};
-
-struct SelectedDownloadProbe {
-  std::string rid;
-  std::string modeName;
-};
-
-// Selects the best probe download candidate using the same matching rank used by downloads.
-static SelectedDownloadProbe SelectBestDownloadCandidateProbe(
-    const std::string &requestedFixtureName,
-    const std::string &requestedMode,
-    int requestedFootprint,
-    const std::vector<DownloadCandidateProbe> &candidates) {
-  SelectedDownloadProbe selected;
-  matching::DownloadCandidateRank bestRank;
-  for (const auto &candidate : candidates) {
-    const auto nameTier = matching::ComputeFixtureNameMatchTier(candidate.fixtureName,
-                                                               requestedFixtureName);
-    if (nameTier == matching::FixtureNameMatchTier::None)
-      continue;
-
-    const auto modeScore = matching::ComputeGdtfModeMatchScore(
-        requestedMode, candidate.modes, requestedFootprint);
-    matching::DownloadCandidateRank rank{nameTier, modeScore.footprintMatch, false,
-                                         candidate.recency, candidate.rating};
-    rank.modeTier = modeScore.tier;
-    if (!matching::IsBetterDownloadCandidate(rank, bestRank))
-      continue;
-
-    bestRank = rank;
-    selected = {candidate.rid, modeScore.modeName};
-  }
-  return selected;
+  assert(catalog::IsBetterDownloadCandidate(manufacturerMatch,
+                                            newerWithoutManufacturer));
+  assert(!catalog::IsBetterDownloadCandidate(newerWithoutManufacturer,
+                                             manufacturerMatch));
 }
 
 // Verifies 12CH digit signatures keep Standard mode requests aligned with mode-aware fixtures.
 static void VerifyStandardModeDigitSignatureBeatsNewerFootprintOnlyMatch() {
-  const std::vector<DownloadCandidateProbe> candidates = {
-      {"fixture-standard", "BLED", {{"Basic", 8}, {"Touring 12CH", 24}}, 100, 3.0f},
-      {"fixture-newer", "BLED", {{"Economy", 12}}, 200, 5.0f},
+  const std::vector<catalog::GdtfCatalogEntry> candidates = {
+      {"fixture-standard", "", "BLED", {{"Basic", 8}, {"Touring 12CH", 24}}, 100, 3.0f},
+      {"fixture-newer", "", "BLED", {{"Economy", 12}}, 200, 5.0f},
   };
 
-  const auto selected = SelectBestDownloadCandidateProbe(
-      "BLED", "Standard mode 12CH", 12, candidates);
+  const auto selected = catalog::SelectBestDownloadMatch(
+      "BLED", "", "Standard mode 12CH", "", 12, candidates);
 
   assert(selected.rid == "fixture-standard");
   assert(selected.modeName == "Touring 12CH");
+  assert(selected.selectionReason == "name+mode-digits");
 }
 
 // Verifies simple named modes without digits outrank footprint-only alternatives.
 static void VerifyBasicModeNameBeatsFootprintOnlyMatch() {
-  const std::vector<DownloadCandidateProbe> candidates = {
-      {"fixture-basic", "Wash 600", {{"Basic", 16}, {"Extended", 24}}, 100, 3.0f},
-      {"fixture-footprint", "Wash 600", {{"Arc Layout", 16}}, 250, 5.0f},
+  const std::vector<catalog::GdtfCatalogEntry> candidates = {
+      {"fixture-basic", "", "Wash 600", {{"Basic", 16}, {"Extended", 24}}, 100, 3.0f},
+      {"fixture-footprint", "", "Wash 600", {{"Arc Layout", 16}}, 250, 5.0f},
   };
 
-  const auto selected = SelectBestDownloadCandidateProbe(
-      "Wash 600", "Basic", 16, candidates);
+  const auto selected = catalog::SelectBestDownloadMatch(
+      "Wash 600", "", "Basic", "", 16, candidates);
 
   assert(selected.rid == "fixture-basic");
   assert(selected.modeName == "Basic");
+  assert(selected.selectionReason == "name+mode");
 }
 
 // Verifies fixture-specific named modes stay paired with the matching catalog mode.
 static void VerifyFixtureSpecificNamedModeStaysAligned() {
-  const std::vector<DownloadCandidateProbe> candidates = {
-      {"fixture-shapes", "Aleda K10 B-EYE", {{"B-EYE Shapes", 35}, {"Standard", 21}}, 100, 3.0f},
-      {"fixture-digits", "Aleda K10 B-EYE", {{"Pixel 35CH", 35}, {"Standard", 21}}, 300, 5.0f},
+  const std::vector<catalog::GdtfCatalogEntry> candidates = {
+      {"fixture-shapes", "", "Aleda K10 B-EYE", {{"B-EYE Shapes", 35}, {"Standard", 21}}, 100, 3.0f},
+      {"fixture-digits", "", "Aleda K10 B-EYE", {{"Pixel 35CH", 35}, {"Standard", 21}}, 300, 5.0f},
   };
 
-  const auto selected = SelectBestDownloadCandidateProbe(
-      "Aleda K10 B-EYE", "B-EYE Shapes", 35, candidates);
+  const auto selected = catalog::SelectBestDownloadMatch(
+      "Aleda K10 B-EYE", "", "B-EYE Shapes", "", 35, candidates);
 
   assert(selected.rid == "fixture-shapes");
   assert(selected.modeName == "B-EYE Shapes");
+  assert(selected.selectionReason == "name+mode");
 }
 
-// Runs focused coverage for MVR-requested GDTF import matching identity selection.
+// Verifies manufacturer evidence is applied during final catalog selection.
+static void VerifyFinalSelectionUsesManufacturerTieBreaker() {
+  const std::vector<catalog::GdtfCatalogEntry> candidates = {
+      {"wrong-brand", "Other Lighting", "My Beam 200", {{"Basic", 16}}, 300, 5.0f},
+      {"right-brand", "Acme GmbH", "My Beam 200", {{"Basic", 16}}, 100, 3.0f},
+  };
+
+  const auto selected = catalog::SelectBestDownloadMatch(
+      "My Beam 200", "", "Basic", "Acme", 16, candidates);
+
+  assert(selected.rid == "right-brand");
+  assert(selected.modeName == "Basic");
+  assert(selected.selectionReason == "name+mode");
+}
+
+// Runs focused coverage for MVR-requested GDTF import and catalog matching selection.
 int main() {
   VerifyDistinctOriginalNamesBeatSharedPlaceholderMetadata();
   VerifySpecBasenameFallback();
@@ -180,5 +154,6 @@ int main() {
   VerifyStandardModeDigitSignatureBeatsNewerFootprintOnlyMatch();
   VerifyBasicModeNameBeatsFootprintOnlyMatch();
   VerifyFixtureSpecificNamedModeStaysAligned();
+  VerifyFinalSelectionUsesManufacturerTieBreaker();
   return 0;
 }
